More code review, e.g. use Tcl_SetObjResult in stead of Tcl_SetResult, preventing a (char *) type case. No functional changes.

+++ generic/tclInt.h
@@ -4875,7 +4875,7 @@ typedef struct NRE_callback {
  * Other externals.
  */
 
-MODULE_SCOPE unsigned long TclEnvEpoch; /* Epoch of the tcl environment 
+MODULE_SCOPE size_t TclEnvEpoch; /* Epoch of the tcl environment
                                          * (if changed with tcl-env). */
 
 #endif /* _TCLINT */
